code optimize

+++ lib/layouts/pages/profile-pages/profile/user_profile_page.dart
@@ -219,8 +219,8 @@ class _UserProfilePageState extends State<UserProfilePage>
                     hCardApiData: widget.userProfileData,
                   ),
                   Container(
-                    height: SS.sH(context) * 0.065 as double,
-                    width: SS.sW(context) as double,
+                    height: SS.sH(context) * 0.065,
+                    width: SS.sW(context),
                     decoration: BoxDecoration(
                       color: ColorPicker.kWhite,
                       border: Border(
@@ -229,7 +229,7 @@ class _UserProfilePageState extends State<UserProfilePage>
                     ),
                     child: Padding(
                       padding: EdgeInsets.symmetric(
-                          horizontal: SS.sW(context) * .080 as double),
+                          horizontal: SS.sW(context) * .080),
                       child: Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
@@ -331,7 +331,7 @@ class _UserProfilePageState extends State<UserProfilePage>
                     ),
                   ),
                   SizedBox(
-                    height: SS.sH(context) * 1.9 as double,
+                    height: SS.sH(context) * 1.9,
                     child: TabBarView(
                         controller: _tabController,
                         children: <Widget>[
